Clean up BrowserBar and update tests for it
