+++ src/ensembl/src/content/app/browser/browser-bar/BrowserBar.tsx
@@ -1,32 +1,12 @@
 import React from 'react';
 import { connect } from 'react-redux';
-import classNames from 'classnames';
-
-import { TrackSet } from '../track-panel/trackPanelConfig';
-import { BreakpointWidth } from 'src/global/globalConfig';
 
 import {
-  getBrowserNavOpened,
   getChrLocation,
   getDefaultChrLocation,
-  getBrowserActivated,
-  getBrowserActiveGenomeId,
   getBrowserActiveEnsObject
 } from '../browserSelectors';
 import { getIsDrawerOpened } from '../drawer/drawerSelectors';
-import {
-  getSelectedTrackPanelTab,
-  getIsTrackPanelModalOpened,
-  getIsTrackPanelOpened
-} from '../track-panel/trackPanelSelectors';
-import { getBreakpointWidth } from 'src/global/globalSelectors';
-
-import { toggleBrowserNav, changeFocusObject } from '../browserActions';
-import {
-  selectTrackPanelTab,
-  toggleTrackPanel
-} from '../track-panel/trackPanelActions';
-import { closeDrawer } from '../drawer/drawerActions';
 
 import BrowserReset from '../browser-reset/BrowserReset';
 import {
@@ -42,23 +22,10 @@ import { EnsObject } from 'src/shared/state/ens-object/ensObjectTypes';
 import styles from './BrowserBar.scss';
 
 export type BrowserBarProps = {
-  // activeGenomeId: string | null;
-  // breakpointWidth: BreakpointWidth;
-  // browserActivated: boolean;
-  // browserNavOpened: boolean;
   chrLocation: ChrLocation | null;
   defaultChrLocation: ChrLocation | null;
   isDrawerOpened: boolean;
-  // isTrackPanelModalOpened: boolean;
-  // isTrackPanelOpened: boolean;
   ensObject: EnsObject | null;
-  // selectedTrackPanelTab: TrackSet;
-  // isFocusObjectInDefaultPosition: boolean;
-  // closeDrawer: () => void;
-  // selectTrackPanelTab: (selectedTrackPanelTab: TrackSet) => void;
-  toggleBrowserNav: () => void;
-  // toggleTrackPanel: (isTrackPanelOpened: boolean) => void;
-  // changeFocusObject: (objectId: string) => void;
 };
 
 type BrowserInfoProps = {
@@ -67,6 +34,7 @@ type BrowserInfoProps = {
 };
 
 export const BrowserBar = (props: BrowserBarProps) => {
+  // FIXME: is this still necessary for anything?
   const shouldShowBrowserInfo = () => {
     const { defaultChrLocation } = props;
     const isLocationOfWholeChromosome = !defaultChrLocation;
@@ -74,14 +42,8 @@ export const BrowserBar = (props: BrowserBarProps) => {
     return !isLocationOfWholeChromosome;
   };
 
-  const browserInfoClassName = classNames(styles.browserInfo, {
-    // [styles.browserInfoGreyed]: isDrawerOpened
-  });
-
-  // const browserRegionClassName = classNames(styles.browserInfoRegion, {
-  //   [styles.browserInfoHidden]: isDrawerOpened
-  // });
-
+  // return empty div instead of null, so that the dedicated slot in the CSS grid of StandardAppLayout
+  // always contains a child DOM element
   if (!(props.chrLocation && props.ensObject)) {
     return <div />;
   }
@@ -102,23 +64,6 @@ export const BrowserBar = (props: BrowserBarProps) => {
   );
 };
 
-/*
-
-{shouldShowTrackPanelTabs && (
-  <TrackPanelTabs
-    closeDrawer={props.closeDrawer}
-    ensObject={props.ensObject}
-    isDrawerOpened={props.isDrawerOpened}
-    selectTrackPanelTab={props.selectTrackPanelTab}
-    selectedTrackPanelTab={props.selectedTrackPanelTab}
-    toggleTrackPanel={props.toggleTrackPanel}
-    isTrackPanelModalOpened={props.isTrackPanelModalOpened}
-    isTrackPanelOpened={props.isTrackPanelOpened}
-  />
-)}
-
-*/
-
 export const BrowserInfo = (props: BrowserInfoProps) => {
   const { ensObject, isDrawerOpened } = props;
   const childProps = {
@@ -135,25 +80,10 @@ export const BrowserInfo = (props: BrowserInfoProps) => {
 };
 
 const mapStateToProps = (state: RootState) => ({
-  activeGenomeId: getBrowserActiveGenomeId(state),
-  breakpointWidth: getBreakpointWidth(state),
-  browserActivated: getBrowserActivated(state),
-  browserNavOpened: getBrowserNavOpened(state),
   chrLocation: getChrLocation(state),
   defaultChrLocation: getDefaultChrLocation(state),
   ensObject: getBrowserActiveEnsObject(state),
-  isDrawerOpened: getIsDrawerOpened(state),
-  isTrackPanelModalOpened: getIsTrackPanelModalOpened(state),
-  isTrackPanelOpened: getIsTrackPanelOpened(state),
-  selectedTrackPanelTab: getSelectedTrackPanelTab(state)
+  isDrawerOpened: getIsDrawerOpened(state)
 });
 
-const mapDispatchToProps = {
-  closeDrawer,
-  selectTrackPanelTab,
-  toggleBrowserNav,
-  toggleTrackPanel,
-  changeFocusObject
-};
-
-export default connect(mapStateToProps, mapDispatchToProps)(BrowserBar);
+export default connect(mapStateToProps)(BrowserBar);
